PSA: allow procMount type Unmasked in baseline
a masked proc mount has traditionally been used to prevent untrusted containers from accessing leaky kernel APIs.
However, within a user namespace, typical ID checks protect better than masked proc. Further, allowing unmasked proc
with a user namespace gives access to a container mounting sub procs, which opens avenues for container-in-container use cases.

Update PSS for baseline to allow a container to access an unmasked /proc, if it's in a user namespace. Update the restricted check
to continue to block with the old behavior (I personally am not positive we need to do this, but as a first step it seems prudent, just in case).

Signed-off-by: Peter Hunt <pehunt@redhat.com>